+++ check_utils.py
@@ -1,0 +1,3 @@
+from env_utils import settings
+
+print(settings)
