+++ mongodb/mongo_setup.py
@@ -1,20 +1,4 @@
 from mongoengine import connect, disconnect
-from pydantic_settings import BaseSettings
-from dotenv import load_dotenv
-
-load_dotenv()
-
-
-# Define MongoDB settings class
-class MongoDBSettings(BaseSettings):
-    mongo_client: str
-    mongo_db_name: str
-    mongo_patrons_collection: str
-    mongo_items_collection: str
-    mongo_bills_collection: str
-
-    class Config:
-        env_file = '.env'
 
 
 # Connect to MongoDB function
